+++ src/Data/Functor/Plus.hs
@@ -16,7 +16,7 @@
 ----------------------------------------------------------------------------
 module Data.Functor.Plus
   ( Plus(..)
-  , asum
+  , psum
   , module Data.Functor.Alt
   ) where
 
@@ -24,10 +24,8 @@ import Control.Applicative hiding (some, many)
 import Control.Applicative.Backwards
 import Control.Applicative.Lift
 import Control.Arrow
--- import Control.Exception
 import Control.Monad
 import Control.Monad.Trans.Identity
--- import Control.Monad.Trans.Cont
 import Control.Monad.Trans.Except
 import Control.Monad.Trans.Maybe
 import Control.Monad.Trans.Reader
@@ -46,7 +44,7 @@ import Data.Functor.Product
 import Data.Functor.Reverse
 import qualified Data.Monoid as Monoid
 import Data.Semigroup hiding (Product)
-import Prelude hiding (id, (.))
+import Prelude hiding (id, (.), foldr)
 
 #if !(MIN_VERSION_transformers(0,6,0))
 import Control.Monad.Trans.Error
@@ -83,16 +81,17 @@ import GHC.Generics
 -- > m <!> zero = m
 --
 -- If extended to an 'Alternative' then 'zero' should equal 'empty'.
-
 class Alt f => Plus f where
   zero :: f a
 
 -- | The sum of a collection of actions, generalizing 'concat'.
 --
--- asum [Just "Hello", Nothing, Just "World"]
--- Just "Hello"
-asum :: (Foldable t, Plus f) => t (f a) -> f a
-asum = foldr (<!>) zero
+-- > asum [Just "Hello", Nothing, Just "World"]
+-- >   -- Just "Hello"
+--
+-- @since 5.3.6
+psum :: (Foldable t, Plus f) => t (f a) -> f a
+psum = foldr (<!>) zero
 
 instance Plus Proxy where
   zero = Proxy